refactor: App to enums

Took 9 minutes
